add small working
add small worjing

+++ src/Graph/graph.h
@@ -37,7 +37,7 @@ public:
 
     void setLongitude(double longitude);
 
-    Edge* addEdge(Vertex* dest, int distance);
+    Edge* addEdge(Vertex* dest, double distance);
     const string &getLabel() const;
 
     void setLabel(const string &label);
@@ -61,12 +61,12 @@ protected:
 
 class Edge {
 public:
-    Edge(Vertex* orig, Vertex* dest, int distance);
-    void setDistance(int distance);
+    Edge(Vertex* orig, Vertex* dest, double distance);
+    void setDistance(double distance);
     ~Edge();
 
     Vertex* getDest();
-    int getdistance() const;
+    double getdistance() const;
     bool isSelected() const; 
     Vertex* getOrig() const; 
     Edge* getReverse() const; 
@@ -74,7 +74,7 @@ public:
     void setReverse(Edge* reverse); 
 protected:
     Vertex* dest; 
-    int distance;
+    double distance;
     bool selected = false;
     Vertex* orig; 
     Edge* reverse = nullptr; 
@@ -92,7 +92,7 @@ public:
 
     bool removeVertex(const string &code);
 
-    bool addEdge(const string &source, const string &dest, int distance, bool direction);
+    bool addEdge(const string &source, const string &dest, double distance, bool direction);
 
     bool removeEdge(const string &source, const string &dest);
